move files to /entities subpackage
